+++ changelog.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- New button to view attendances-report without download.
+
 ## [1.3.1] - 2020-08-18
 
 ### Added